requirements.txt
langchain
langchain-community
fastapi
uvicorn
chromadb
sentence-transformers
python-dotenv
python-multipart